backend/requirements.txt
argcomplete==1.10.3
asgiref==3.8.1
beautifulsoup4==4.8.2
cachetools==5.4.0
certifi==2024.7.4
channels==4.1.0
chardet==3.0.4
charset-normalizer==3.3.2
compressed-rtf==1.0.6
decorator==4.4.2
Django==5.0.7
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docx2txt==0.8
ebcdic==1.1.1
extract-msg==0.28.7
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-speech==2.26.1
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.65.1
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
IMAPClient==2.1.0
lxml==5.2.2
moviepy==1.0.3
numpy==2.0.1
olefile==0.47
pdfminer.six==20191110
pillow==10.4.0
proglog==0.1.10
proto-plus==1.24.0
protobuf==5.27.2
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodome==3.20.0
pydub==0.25.1
PyJWT==2.8.0
python-pptx==0.6.23
pytz==2024.1
requests==2.32.3
rsa==4.9
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
sqlparse==0.5.0
stripe==2.76.0
textract==1.6.5
tqdm==4.66.4
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
xlrd==1.2.0
XlsxWriter==3.2.0
